requirements/dev.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile dev.in
#
alabaster==0.7.12
    # via sphinx
attrs==21.4.0
    # via pytest
babel==2.10.3
    # via sphinx
black==22.6.0
    # via -r dev.in
build==0.8.0
    # via
    #   check-manifest
    #   pip-tools
certifi==2022.6.15
    # via requests
charset-normalizer==2.1.0
    # via requests
check-manifest==0.48
    # via -r check-manifest.in
click==8.1.3
    # via
    #   black
    #   docstr-coverage
    #   flask
    #   pip-tools
coverage[toml]==6.4.2
    # via
    #   -r coverage.in
    #   pytest-cov
distlib==0.3.5
    # via virtualenv
docstr-coverage==2.2.0
    # via -r docstr-cov.in
docutils==0.16
    # via
    #   sphinx
    #   sphinx-tabs
filelock==3.7.1
    # via
    #   tox
    #   virtualenv
flake8==4.0.1
    # via
    #   -r style.in
    #   flake8-builtins
    #   pep8-naming
flake8-builtins==1.5.3
    # via -r style.in
flake8-import-order==0.18.1
    # via -r style.in
flask==2.0.1
    # via -r sample_app.in
idna==3.3
    # via requests
imagesize==1.4.1
    # via sphinx
iniconfig==1.1.1
    # via pytest
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via
    #   flask
    #   sphinx
markupsafe==2.1.1
    # via jinja2
mccabe==0.6.1
    # via flake8
mypy-extensions==0.4.3
    # via black
packaging==21.3
    # via
    #   build
    #   pytest
    #   sphinx
    #   tox
pathspec==0.9.0
    # via black
pep517==0.12.0
    # via build
pep8-naming==0.13.1
    # via -r style.in
pip-tools==6.8.0
    # via -r dev.in
platformdirs==2.5.2
    # via
    #   black
    #   virtualenv
pluggy==1.0.0
    # via
    #   pytest
    #   tox
py==1.11.0
    # via
    #   pytest
    #   tox
pycodestyle==2.8.0
    # via
    #   flake8
    #   flake8-import-order
pyflakes==2.4.0
    # via flake8
pygments==2.12.0
    # via
    #   sphinx
    #   sphinx-tabs
pyparsing==3.0.9
    # via packaging
pytest==7.1.2
    # via
    #   -r tests.in
    #   pytest-cov
pytest-cov==3.0.0
    # via -r coverage.in
python-dotenv==0.21.1
    # via -r sample_app.in
pytz==2022.1
    # via babel
pyyaml==6.0
    # via docstr-coverage
requests==2.28.1
    # via
    #   -r sample_app.in
    #   sphinx
six==1.16.0
    # via
    #   tox
    #   virtualenv
snowballstemmer==2.2.0
    # via sphinx
sphinx==4.2.0
    # via
    #   -r docs.in
    #   sphinx-tabs
sphinx-tabs==3.2.0
    # via -r docs.in
sphinxcontrib-applehelp==1.0.2
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==2.0.0
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via sphinx
toml==0.10.2
    # via tox
tomli==2.0.1
    # via
    #   black
    #   build
    #   check-manifest
    #   coverage
    #   pytest
tox==3.25.1
    # via -r dev.in
urllib3==1.26.10
    # via requests
virtualenv==20.15.1
    # via tox
werkzeug==2.0.3
    # via
    #   -r sample_app.in
    #   flask
wheel==0.37.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
